refactor(work): extract RED load-failure heuristic to shared lib
Extract RED_LOAD_FAILURE_PATTERNS and related logic to new plugins/work/scripts/workflows/work-implement/lib/red-load-failure.js so future consumers (enforce-tdd-on-stop.js) can reuse without copy-paste drift. tdd-phase-state.js now requires from the shared lib. Add subsection to tdd-enforcement.md documenting rejection condition and recovery path.
